src/main/java/pl/kkowalewski/recipeapp/converter/tocommand/UnitOfMeasureToUnitOfMeasureCommand.java
package pl.kkowalewski.recipeapp.converter.tocommand;

import lombok.Synchronized;
import org.springframework.core.convert.converter.Converter;
import org.springframework.lang.Nullable;
import org.springframework.stereotype.Component;
import pl.kkowalewski.recipeapp.command.UnitOfMeasureCommand;
import pl.kkowalewski.recipeapp.converter.Command;
import pl.kkowalewski.recipeapp.model.UnitOfMeasure;

@Component
public class UnitOfMeasureToUnitOfMeasureCommand implements Converter<UnitOfMeasure,
        UnitOfMeasureCommand>, Command<UnitOfMeasure, UnitOfMeasureCommand> {

    /*------------------------ FIELDS REGION ------------------------*/

    /*------------------------ METHODS REGION ------------------------*/
    @Synchronized
    @Nullable
    @Override
    public UnitOfMeasureCommand convert(UnitOfMeasure unit) {
        if (unit == null) {
            return null;
        }

        return new UnitOfMeasureCommand(unit.getId(), unit.getDescription());
    }
}
